added ProgramAlgebra identity, fixed removeUnaryIfs
In particular, looking at the [ProgramAlgebra] for [removeUnaryIfs], I
noticed its type was extremely generic, saying it takes any language
with [If1] and returns any other language that supports [If2] and
[Unit].

Because these are loose requirements, this typeclass instance could be
picked up in any operation between such two languages.  We don't want
this to happen, so we add a [Label] to every [ProgramAlgebra], so that
the typeclass resolution mechanism only looks for instances with the
appropriate [Label] for the operation being resolved.

Also, to implement [removeUnaryIfs], I had to define an overlapping instance
that catches all cases but [If1], so that we don't need to write one instance
per extension of the input language.  Pretty neat!